Moves WildFire source and header files to /src and /include
